Development Commit
This patch adds a central widget to the main menu and sets it layout to
a QHBoxLayout. Then, a video widget is created and added onto the
central widget.

This patch also updates the naming convention of class data members to
begin with an '_' character to denote that its private.

Note: On my system, running Fedora 36, the video widget appears to be
buggy (noticeable glitching in the frames) if the video file is selected
but not played.